+++ cdflib/__init__.py
@@ -5,6 +5,8 @@ from . import cdfread, cdfwrite
 from .epochs import CDFepoch as cdfepoch  # noqa: F401
 
 __all__ = ["CDF"]
+
+
 try:
     from ._version import version as __version__
 except Exception: