+++ doc/changelog.rst
@@ -2,6 +2,11 @@
 Changelog
 =========
 
+1.0.4
+=====
+- Fixed issue where multi-dimensional variables were dropped when converting to xarray.
+- Replaced all print and warning statements with a logger, ``cdflib.logging.logger``.
+
 1.0.3
 =====
 - The ``variable`` parameter to `cdflib.cdfread.CDF.varattsget` is no longer optional.